chore: add vscode launch config
Adds a launch config for the example project.

This allows quickly running the example, for instance with the F5 key,
without having to open the main.dart file first.

This especially makes debugging native code easier, as you don't have
to switch back to the main.dart file to start the example.